updated example test
