+++ package.json
@@ -9,23 +9,29 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.1",
+    "@react-navigation/stack": "^6.0.1",
     "expo": "~42.0.1",
     "expo-status-bar": "~1.0.4",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
-    "react-native-web": "~0.13.12"
+    "react-native-gesture-handler": "^1.10.3",
+    "react-native-web": "~0.13.12",
+    "react-redux": "^7.2.4",
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
     "@types/react": "~16.9.35",
     "@types/react-native": "~0.63.2",
-    "typescript": "~4.0.0",
     "enzyme": "^3.11.0",
     "jest": "^26.6.3",
+    "jest-expo": "^42.1.0",
     "jest-expo-enzyme": "^1.2.1",
+    "redux-mock-store": "^1.5.4",
     "ts-jest": "^26.5.6",
-    "redux-mock-store": "^1.5.4"
+    "typescript": "~4.0.0"
   },
   "private": true
 }
